Avoid including the promise polyfill when possible
When we know we are running on Ember versions that do not support IE11,
we can avoid including the promise polyfill all together.
