+++ package.json
@@ -38,6 +38,7 @@
   },
   "dependencies": {
     "@ember/test-waiters": "^3.0.0",
+    "@embroider/macros": ">= 0.47.0",
     "broccoli-debug": "^0.6.5",
     "broccoli-funnel": "^3.0.8",
     "ember-cli-babel": "^7.26.6",
